fix merge crash of pbxproj file. Apple burn in hell, please

+++ Project/Zazo/UserStory/Main/Grid Module/UserInterface/View/Views/UserRecorderGridView/ZZUserRecorderGridView+AnimationExtension.m
@@ -83,7 +83,7 @@
 - (void)_showVideoCountLabelWithCount:(NSInteger)count
 {
     self.videoCountLabel.hidden = NO;
-    self.videoCountLabel.text = [NSString stringWithFormat:@"%i",count];
+    self.videoCountLabel.text = [NSString stringWithFormat:@"%li",(long)count];
     self.backgroundColor = [UIColor an_colorWithHexString:kLayoutConstGreenColor];
 
 }
